Renamed resolver to enricher
